Added bootstrap services order and changed Hibernate app start action
UserBootstrap service need to be run after RolesBotstrap because users roles are got from database. Without this BootstrapServices could be run in unpredictables orders, which causes that users creates wihout roles and can't get acces to other services.
Hibernate ddl-auto was changed because 'create' or 'create-drop' deletes not all data in database and leaves rows which were created in Bootstrap, but deletes data in connective tables such as 'user_roles', which leaves users without roles